Feature/java21

+++ src/main/java/com/labelzoom/api/model/components/barcodes/IBarcodeWithModuleWidth.java
@@ -32,6 +32,6 @@ public interface IBarcodeWithModuleWidth
 
     default float getModuleWidthInMm(final int dpi)
     {
-        return (float)this.getBarcodeSize().getValue() / (float)dpi * 25.4f;
+        return (float)this.getBarcodeSize().getValue() / dpi * 25.4f;
     }
 }
